pyproject.toml
[project]
name = "dukamarket"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.11.1",
    "celery==5.5.3",
    "django==4.2.16",
    "django-cors-headers==4.9.0",
    "django-import-export==4.1.1",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "pillow==10.4.0",
    "python-decouple==3.8",
    "redis==5.2.1",
    "social-auth-app-django==5.4.0",
    "sqlparse==0.5.5",
    "tzdata==2025.3",
    "urllib3==2.6.3",
]
